+++ doc/release_notes.rst
@@ -4,9 +4,28 @@ Release Notes
 Upcoming Release
 ----------------
 
+
+**New Features**
+
+* It is now possible to set the sense of the objective function to `minimize` or `maximize`. Therefore, a new class `Objective` was introduced which is used in `Model.objective`. It supports the same arithmetic operations as `LinearExpression` and `QuadraticExpression` and contains a `sense` attribute which can be set to `minimize` or `maximize`.
+v* The `fillna` function for variables was made more secure by raising a warning if the fill value is not of  variable-like type.
+* The `where` and `fillna` functions for expressions were made more flexible: When passing a scalar value or a DataArray, the values are added as constants to the expression, where there were missing values before. If another expression is passed, the values are added to the expression, where there were missing values before.
+* New solver interface for `SCIP <https://www.scipopt.org/>`. This solver is now supported by `linopy` and can be used with the `solve` function if the `pyscipopt` package is installed. The solver is available for free for general use. See the `SCIP website <https://www.scipopt.org/>` for more information.
+
+**Breaking Changes**
+
+* The `_fill_value` for LinearExpression and QuadraticExpression classes was changed to ``NaN`` for the constant array ("const"). This allows to use the `where` function for expressions with constant values in the argument `other`.
+* The functions ``ravel`` and ``iter_ravel`` for Variables and Constraints were removed in favor of the ``flat`` function.
+* The property ``non_helper_dims`` for Variables and Constraints was removed in favor of the ``coord_dims`` property.
+* The function ``to_anonymous_constraint`` was removed in favor of the ``to_constraint`` function.
+
+
+Version 0.2.5
+-------------
+
 * The memory-efficiency of the IO to LP/MPS file was further improved. In particular, the function `to_dataframe` is now avoiding unnecessary data copies.
 * The printout of time stamps was modified to be more readable, leaving out the display of seconds and below if not necessary.
-* New solver interface for `SCIP <https://www.scipopt.org/>`. This solver is now supported by `linopy` and can be used with the `solve` function if the `pyscipopt` package is installed. The solver is available for free for general use. See the `SCIP website <https://www.scipopt.org/>` for more information.
+* The gurobi environment is now enclosed in a context manager to avoid any unwanted use of a token.
 
 
 Version 0.2.5